changes for postgres

+++ data/migrations/20210819034318_create_users.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
       .notNullable();
       tbl.string('password')
       .notNullable();
-      tbl.integer('phone_number');
+      tbl.bigint('phone_number');
   });
 };
